+++ src/main/webapp/resources/js/components/ant.design/table-renderers.js
@@ -16,6 +16,7 @@ export const nameColumnFormat = ({ url }) => {
     width: 200,
     sorter: true,
     ellipsis: true,
+    className: "t-name-col",
     filterIcon(filtered) {
       return <IconTableFilter className="t-name" filtered={filtered} />;
     },
@@ -29,8 +30,10 @@ export const nameColumnFormat = ({ url }) => {
   };
 };
 
-export const dateColumnFormat = () => ({
+export const dateColumnFormat = ({ className = "" } = {}) => ({
   sorter: true,
   width: 230,
-  render: date => formatInternationalizedDateTime(date)
+  render: date => (
+    <span className={className}>{formatInternationalizedDateTime(date)}</span>
+  )
 });