+++ src/main/webapp/resources/js/pages/projects/samples/linker/Linker.jsx
@@ -1,12 +1,13 @@
-import React from "react";
+import React, { useEffect, useState } from "react";
 import { render } from "react-dom";
-import { Button, Modal, Typography } from "antd";
-import { CodeOutlined } from "@ant-design/icons";
+import { Button, Checkbox, Form, Modal, Typography } from "antd";
+
 import { grey2, grey9 } from "../../../../styles/colors";
 import { getNGSLinkerCode } from "../../../../apis/linker/linker";
 import { SPACE_SM } from "../../../../styles/spacing";
 import { BORDER_RADIUS, BORDERED_LIGHT } from "../../../../styles/borders";
 import styled from "styled-components";
+import { IconCode } from "../../../../components/icons/Icons";
 
 const { Paragraph, Text } = Typography;
 
@@ -27,32 +28,29 @@ const CommandText = styled(Paragraph)`
  * @constructor
  */
 function Linker() {
-  function handleSampleIds(e) {
-    // Post data to get linker command.
-    const { detail } = e;
+  const [visible, setVisible] = useState(false);
+  const [types, setTypes] = useState(["fastq"]);
+  const [scriptString, setScriptString] = useState();
+  const [command, setCommand] = useState();
+
+  const options = [
+    { label: i18n("Linker.fastq"), value: "fastq" },
+    { label: i18n("Linker.assembly"), value: "assembly" }
+  ];
+
+  function handleSampleIds({ detail }) {
     getNGSLinkerCode(detail).then(({ data }) => {
-      Modal.success({
-        className: "t-linker-modal",
-        width: 500,
-        title: i18n("Linker.title"),
-        content: (
-          <>
-            <Paragraph>{i18n("Linker.details")}</Paragraph>
-            <Text type="secondary">
-              <span dangerouslySetInnerHTML={{ __html: i18n("Linker.note") }} />
-            </Text>
-            <CommandText
-              className="t-cmd-text"
-              ellipsis={{ rows: 1 }}
-              copyable={{ text: data }}
-            >
-              {data}
-            </CommandText>
-          </>
-        )
-      });
+      // Post data to the server to get the linker command.
+      setScriptString(data);
+      setVisible(true);
     });
   }
+
+  useEffect(() => {
+    setCommand(
+      types.length ? `${scriptString} -t ${types.join(",")}` : scriptString
+    );
+  }, [scriptString, types]);
 
   /*
   These Listeners and Dispatchers are a way to get around the separation between react
@@ -60,25 +58,55 @@ function Linker() {
   */
   document.addEventListener("sample-ids-return", handleSampleIds, false);
 
-  function showModal() {
+  function getIds() {
     document.dispatchEvent(new Event("sample-ids"));
   }
 
   return (
-    <Button
-      type="link"
-      style={{
-        margin: `0 inherit`,
-        padding: 0,
-        color: grey9
-      }}
-      className="t-linker-btn"
-      onClick={showModal}
-    >
-      {/* constant marginRight here to match old styles */}
-      <CodeOutlined style={{ marginRight: 2 }} />
-      {i18n("project.samples.export.linker")}
-    </Button>
+    <>
+      <Button
+        type="link"
+        style={{
+          margin: `0 inherit`,
+          padding: 0,
+          color: grey9
+        }}
+        className="t-linker-btn"
+        onClick={getIds}
+      >
+        {/* constant marginRight here to match old styles */}
+        <IconCode style={{ marginRight: 2 }} />
+        {i18n("project.samples.export.linker")}
+      </Button>
+      <Modal
+        title={i18n("Linker.title")}
+        className="t-linker-modal"
+        visible={visible}
+        footer={null}
+        onCancel={() => setVisible(false)}
+      >
+        <Paragraph>{i18n("Linker.details")}</Paragraph>
+        <Text type="secondary">
+          <span dangerouslySetInnerHTML={{ __html: i18n("Linker.note") }} />
+        </Text>
+        <Form layout="inline">
+          <Form.Item label={i18n("Linker.types")}>
+            <Checkbox.Group
+              options={options}
+              defaultValue={["fastq"]}
+              onChange={setTypes}
+            />
+          </Form.Item>
+        </Form>
+        <CommandText
+          className="t-cmd-text"
+          ellipsis={{ rows: 1 }}
+          copyable={command}
+        >
+          {command}
+        </CommandText>
+      </Modal>
+    </>
   );
 }