+++ src/main/webapp/resources/js/utilities/url-utilities.js
@@ -10,14 +10,14 @@
  */
 export function setBaseUrl(url) {
   /*
-  Get the base url which is set via the thymeleaf template engine.
-   */
+    Get the base url which is set via the thymeleaf template engine.
+     */
   const BASE_URL = window.TL?._BASE_URL || "/";
 
   /*
-  Check to make sure that the given url has not already been given the
-  base url.
-   */
+    Check to make sure that the given url has not already been given the
+    base url.
+     */
   if (url.startsWith(BASE_URL) || url.startsWith("http")) {
     return url;
   }
@@ -26,7 +26,7 @@ export function setBaseUrl(url) {
   url = url.replace(/^\/+/, "");
 
   /*
-  Create the new url
-   */
+    Create the new url
+     */
   return `${BASE_URL}${url}`;
 }